fix: moved AccountDisplay to own comp
