+++ src/components/Notification/NotificationBanner.css
@@ -13,6 +13,7 @@
 .notification-banner p {
     font-size: 16px;
     font-weight: 400;
+    font-style: italic;
 }
 
 .notification-banner .btn {
@@ -25,9 +26,7 @@
 }
 
 .toast-message {
-    background: rgb(35, 31, 32);
-    color: #f6f6f6;
+    background: rgba(12, 12, 12, 0.9);
+    color: #ffffff;
     font-size: 20px;
-    width: 34vw;
-    padding: 16px
 }
