+++ src/components/ui/AccountDisplay/AccountDisplay.css
@@ -1,0 +1,6 @@
+.account-display {
+    display: 'flex';
+    align-items: 'center';
+    gap: '12px';
+    font-weight: 500;
+}
